@@ -2,18 +2,18 @@ package guru.springframework.converters;
 
 import guru.springframework.commands.CategoryCommand;
 import guru.springframework.domain.Category;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class CategoryCommandToCategoryTest {
 
-    public static final Long ID_VALUE = new Long(1L);
+    public static final Long ID_VALUE =  Long.valueOf(1L);
     public static final String DESCRIPTION = "description";
     CategoryCommandToCategory conveter;
 
-    @Before
+    @BeforeEach
     public void setUp() throws Exception {
         conveter = new CategoryCommandToCategory();
     }
